+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Added
 - [MINOR] Added `background` prop to `KibaApp`
+- [MINOR] Created `renderHead` for external head rendering
 
 ### Changed
 - [MINOR] Fixed colors theme creation to calculate text and textOnBrand correctly when background is dark